fix(nvim): resolve startup errors due to lazy loading plugins
I was having inconsistent issues with LSP plugins and Telescope
depending on how neovim started up. null-ls was failing to attach
to the buffer because it was trying before the lsp client was
initialized. The keymaps defined in which-key were also
inconsistently available. I made a couple other lazy loading tweaks
as well which should help.